re #757: Optimized speed of vtkOrientedImageDataResample::CalculateEffectiveExtent
CalculateEffectiveExtent is still the largest contributor to slow response time at effect application for large volumes.
It might be possible to further optimize the code but most probably the best would be to allow effects to provide what extent they updated. It would be optional and would only be used by paint, draw, erase, etc. effects (that operate in a small region only).


git-svn-id: https://subversion.assembla.com/svn/slicerrt/trunk/SlicerRt/src@2725 25a7314f-c4c8-4314-9936-a25f8480a772