re #757: Reworked segment editor
The edited labelmap is just a modifier image now (containing the last stroke that has to be added/removed/replaced in the existing labelmap). This allows applying the modification to the current segment and its inverse to other segments. This enables a new key feature: masking. Masking can be used to restrict editing to certain regions (inside/outside specific segments) and to control overlap/overpaint.

Also reworked how segment labelmap geometry is determined. Geometry defined in "reference image geometry" conversion parameter is used as edited segment labelmap geometry. If it's not defined then geometry of existing segments are used. If there were no existing segments either, then the master volume's geometry is used. If master volume is changed, segment labelmap geometry remains the same and so other various volumes can be used as master volumes (they are resampled to match the segment labelmap geometry).

Painting can now be performed on any slice orientations, as a 3D brush is generated. The 3D brush is shown in all other views (both 2D and 3D). 2D brush thickness is determined by slice spacing.

Painting is now possible on 3D views, too.

Editable intensity range is now available for all effects. It controls what area of the image the selected segment is allowed to be changed. It brings "threshold paint" to other effects, such as rectangle, drawing, etc.

Experimentally, use built-in VTK morphological operators for dilation.

git-svn-id: https://subversion.assembla.com/svn/slicerrt/trunk/SlicerRt/src@2809 25a7314f-c4c8-4314-9936-a25f8480a772