re #757: Use Teem's vtkNRRDReader and vtkNRRDReader for binary labelmap file read/write
These reader/writer classes are used throughout Slicer to read 4D images. Data is stored as list of 3D volumes instead of a 4D spatial volume.

A side effect is that a segmentation can now be loaded as a sequence (useful for testing).

Reader for the old file format is kept for now for backward compatibility (can be disabled by undefining or removing parts marked with #ifdef SUPPORT_4D_SPATIAL_NRRD).

Requires latest nightly Slicer, as the NRRD reader threw error when attempted to read a 4D spatial nrrd (legacy segmentation file format).


git-svn-id: https://subversion.assembla.com/svn/slicerrt/trunk/SlicerRt/src@2868 25a7314f-c4c8-4314-9936-a25f8480a772