Re #743: Updated External Beam Planning to use Qt-based dose engine plugins instead of VTK-based ones WIP.
This enables adding custom UI for the special beam types, and Python wrapping, which is essential for Matlab-based and Python dose engines and useful for command-line ones. vtkSlicerExternalBeamPlanningModuleLogicTest1 is disabled for now, it will be needed to convert to a Python test. Also the existing dose engines do not work as yet, because the special beam types are not handled with the new plugin mechanism. Thus the module is unfunctional for the moment, but it is experimental, and also fix will arrive in a few days.

git-svn-id: https://subversion.assembla.com/svn/slicerrt/trunk/SlicerRt/src@2982 25a7314f-c4c8-4314-9936-a25f8480a772