BUG: Solved #72, inconsistent vtkPolyDataDistanceHistogramFilter output
After verifying that the distances were computed correctly, I examined the code that computed the histograms from those distances. There was a memcpy call for copying data from a vtkDoubleArray and a vtkImageData. I saw two problems with this:

1) It will not work if the data is not stored in non-continuous memory (data stored across multiple parts of memory).
2) It is not intuitive what is going on.

I rewrote this section of code to be both more correct and more clear.

In addition:
- There is now a regression test for this filter
- Extents used by this filter have been corrected