COMP: Migration to Qt5/VTK8
The module LeakFinder was completely removed, because compared to how useful the module is (it has been used only once as far as I know) and how complex the task of fixing it is (Qt5 requires VS2015 or later, which do not supply one of the libraries StackWalker needs (tchar.h), and verifying that any nontrivial change works would take a long time), it would not be worth the investment. We can always restore the module later if the need arises.