ENH: Allow using multiple isodose sets at the same time
Isodose sets can be created and managed for multiple dose volumes. If the dose volume is changed in the Isodose module, the corresponding isodose set is shown.
This commit also fixes numerous issues in isodose including refactoring of handling of the scalar bar

Fixes https://github.com/SlicerRt/SlicerRT/issues/8