ENH: Practically disable ribbon converter instead of manually excluding it
Instead of sifting through the conversion paths in DVH calculation, it is best to practically disable the planar contours to ribbon filter by increasing its cost to quasi infinity, so that this path is never chosen by default, but the user has to manually request it. This is due to the superiority of the triangulated and end-capped closed surface path.