BUG: Fix beam geometry modification in loaded scene
The beam poly data in beam model node needs to be re-observed after import. Without this, the beam polydata display is not updated when the beam geometry changes. The reason for this is possibly that the pipeline is set up with the file reader and on any modified event that pipeline is used instead of simply using the changed contents of the beam polydata.

Re https://github.com/SlicerRt/SlicerRT/issues/73