BUG: Fix incorrect beam image scalar type for qSlicerMockDoseEngine
The beam was added to the segment and converted from closed surface to scalar representation, however the converted image was not retrieved from the segment after conversion.
The resulted in an error since the scalar type of the labelmap was not unsigned char.

Fixed by retrieving the labelmap from the segment post-conversion.