BUG: Fix shared labelmap bugs in SegmentComparison/SegmentMorphology
In SegmentComparison/SegmentMorphology, the vtkSlicerSegmentationsModuleLogic::GetSegmentBinaryLabelmapRepresentation function was used to get the binary labelmaps. This function returns the internal "shared" representation of the segments, rather than the individual segment labelmap. This caused issues in SegmentComparison/SegmentMorphology, which expected only a single segment in each labelmap.

Fixed by changing vtkSlicerSegmentationsModuleLogic::GetSegmentBinaryLabelmapRepresentation to vtkSegmentationNode::GetBinaryLabelmapRepresentation, which retrieves the individual non-shared labelmap for each representation.