BUG: Fix planar contour to surface contour plane normal calculation
Calculation of plane normal now uses the vtkAddonMathUtilities::FitPlaneToPoints to calculate the correct orientation of the planes.

Plane normal calculation using vtkTextureMapToPlane (not sure why it was used in the first place) was causing issues, so that the transform that aligned the plane normals to the z-axis was failing.

See discussion here: https://discourse.slicer.org/t/planar-contour-to-binary-labelmap-or-model/14828