COMP: Use vtkCollisionDetectionFilter if available in VTK
Previously vtkCollisionDetectionFilter was not part of VTK and therefore it was added to SlicerRT.
The filter was added to VTK at around version 8.90, therefore in these VTK versions we use this built-in filter instead of the SlicerRT version.

The filter is removed completely because even having the header file in the build tree (with a different content than the one in VTK) caused build issues.

The Extension Index is updated to build Slicer-4.11 with the 4.11 branch.

fixes #195