ENH: Make BatchStructureSetConversion more robust
When DICOM data contained the same study instance UID with different patient name/ID then it cause an empty patient appear in the database.
The script stopped with an exception when an empty patient was found.

This commits changes the behavior to just log a message if a patient does not contain data, but it still proceeds with loading the other patients.

fixes https://github.com/Slicer/Slicer/issues/5850