BUG: Fix crash while loading RTDOSE DICOM file
1. Fix crash while loading RTDOSE DICOM file;
2. Fix deprecated Markups API warning;
3. Minor changes.

"Isodose" module logic creates several color tables. In order to change or add data into color table, for isodose and dose visualization,
the color table must be editable or be a copy of a default color table. If the scene was cleared and all the nodes were deleted,
the editable isodose colortable was deleted as well, and it wouldn't be created again with proper lookup table.

For that reason if the user is trying to load RTDOSE volume, the program crashes on RTDOSE export, even if isodose table node pointer is valid!

  Positions in vtkSlicerDicomRtImportExportModuleLogic.cxx file, where crash takes place

  // Set window/level to match the isodose levels
  int minDoseInDefaultIsodoseLevels = vtkVariant(defaultIsodoseColorTable->GetColorName(0)).ToInt();
  int maxDoseInDefaultIsodoseLevels = vtkVariant(defaultIsodoseColorTable->GetColorName(defaultIsodoseColorTable->GetNumberOfColors()-1)).ToInt();

А singleton tag was added to the copy of default isodose color table to fix the crash.

How to test the crash.

1. Load a DICOM series with RTDOSE, for example: https://github.com/SlicerRt/SlicerRtData/tree/master/eclipse-8.1.20-phantom-prostate
2. Clear the scene. "File"->"Close Scene" or Ctrl+W -> "Close Scene"
3. Try to load the same series once again.