BUG: Fix failing vtkSlicerIsodoseModuleLogicTest_EclipseProstate
The isodose surface generated with VTK9 is slightly different around the holes in the bottom of the mesh. This causes a small volume difference between the baseline and generated surface. The surface is the same in all other areas, so there doesn't seem to be an issue. Fixed by updating the baseline.