ENH: Initial support of orientated CT volumes
If IJKToRASDirection matrix _A_ isn't LPS transform,

      | a11 a12 a13 a14 |    | -1  0 0 0 |
_A_ = | a21 a22 a23 a24 | != |  0 -1 0 0 |
      | a31 a32 a33 a34 |    |  0  0 1 0 |
      | a41 a42 a43 a44 |    |  0  0 0 1 |

then, the volume direction to LPS transform matrix _R_ can be applied to the CT volume
to make IJKToRASDirection matrix of the CT volume to be LPS orientated
after hardening the transform.

the matrix transformation:

               | -1  0 0 0 |
_R_ = _A_^-1 * |  0 -1 0 0 |
               |  0  0 1 0 |
               |  0  0 0 1 |

the matrix widget shows elements of the transformation matrix _R_,

if the matrix _A_ is LPS orientated => the matrix _R_ is an identity matrix.