ENH: Let user turn ON/OFF collision detection
In many laptops, the algorithm eats up a lot of RAM and freezes the computer for some seconds. So disable it by default, and only turn it on actively by the user, so that he sees why this is happening.

In some cases, we do not want to calculate collisions, just to play around and show functionality as well as quickly visually see where there might be collisions. Once beam angle is decided, we can activate numerical collision detection, and get more accurate results.