BUG: Prevent crash on startup
There was a crash on startup due to the initialization of the scripted dose engines. This step is now done on entering the External Beam Planning module instead of when setting up the module. It still crashes, but at least the application and the rest of the extension can be used.
Note that there is no crash with Slicer 5.8.1, but only in 5.9

Re #284